// SampleSolution/ClientApp/src/app/dashboard/some-data/models/createDto.ts
export class CreateDto {
  firstName: string;
  middleName: string;
  lastName: string;
  title: string;
  linkedInUrl: string;
  facebookUrl: string;
  twitterUrl: string;
  googlePlusUrl: string;
  skype: string;
  instagramUrl: string;
  snapchat: string;
  youTubeUrl: string;
  color: string;
}